django01/project01/requirements.txt
django-survey-and-report # UNINSTALLED

django-survey

